requirements.txt
sqlalchemy
psycopg2-binary
feedparser
openai
python-dotenv
apscheduler
fastapi
uvicorn
httpx
# For testing purposes
pytest
# For database migrations
alembic
# For asynchronous database operations
asyncpg
# For handling CORS in FastAPI
fastapi-cors
# For handling background tasks in FastAPI
fastapi-background
# For handling JSON Web Tokens
python-jose
# For handling environment variables
python-dotenv